+++ userspace/nvm/types.h
@@ -27,14 +27,14 @@ extern "C" {
  * Note: the queue handle does not "own" the memory buffer it points to.
  * Memory management must be handled elsewhere.
  */
-struct nvm_queue
+struct nvm_queue 
 {
     uint16_t            no;             // Queue number (must be unique per SQ/CQ pair)
-    size_t              max_entries;    // Maximum number of queue entries supported
-    size_t              entry_size;     // The size of a queue entry
+    uint16_t            max_entries;    // Maximum number of queue entries supported
+    uint16_t            entry_size;     // The size of a queue entry
     uint32_t            head;           // Head pointer
     uint32_t            tail;           // Tail pointer
-    int                 phase;          // Current phase bit
+    int16_t             phase;          // Current phase bit
     void*               virt_addr;      // Virtual address of the queue (may be device memory)
     uint64_t            bus_addr;       // Physical/bus address of the queue (only 1 page per queue supported)
     volatile uint32_t*  db;             // Pointer to doorbell register (write only)
